polyglot_backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
Flask==3.1.1
Flask-Cors==6.0.1
Flask-SocketIO==5.5.1
eventlet==0.40.3
Django==5.2.5
dj-database-url==2.2.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.11.7
httpx==0.28.1
requests==2.32.4
gunicorn==23.0.0